+++ Examples/Resources/Solvers/CSCCSat/_sparkle_solver_wrapper.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: UTF-8 -*-
+"""Example CSCCSat Solver wrapper in Python."""
 import sys
 import subprocess
 from pathlib import Path
 from sparkle.types import SolverStatus
-from sparkle.tools.solver_wrapper_parsing import parse_solver_wrapper_args, \
-    get_solver_call_params
+from sparkle.tools.solver_wrapper_parsing import parse_solver_wrapper_args
 
 
 # Parse the arguments of the solver wrapper
@@ -17,19 +17,16 @@ instance = args_dict["instance"]
 seed = args_dict["seed"]
 
 # Construct the base solver call
-solver_name = "PbO-CCSAT"
-solver_exec = f"{solver_dir / solver_name}" if solver_dir != Path(".") else "./" + \
-    solver_name
-solver_cmd = [solver_exec,
-              "-inst", str(instance),
-              "-seed", str(seed)]
+solver_name = "CSCCSat"
+if solver_dir != Path("."):
+    solver_exec = f"{solver_dir / solver_name}"
+else:
+    f"./{solver_name}"
+solver_cmd = [solver_exec, str(instance), str(seed)]
 
-# Get further params for the solver call
-params = get_solver_call_params(args_dict)
-
-# Execute the solver call
+# CSCCSat does not have any configurable parameters, thus other params will not be added
 try:
-    solver_call = subprocess.run(solver_cmd + params,
+    solver_call = subprocess.run(solver_cmd,
                                  capture_output=True)
 except Exception as ex:
     print(f"Solver call failed with exception:\n{ex}")
@@ -38,18 +35,19 @@ except Exception as ex:
 output_str = solver_call.stdout.decode()
 print(output_str)  # Print original output so it can be verified
 
+# Try to parse the status from the output
 status = SolverStatus.CRASHED
-for line in output_str.splitlines():
+for line in reversed(output_str.splitlines()):
     line = line.strip()
-    if (line == r's SATISFIABLE') or (line == r's UNSATISFIABLE'):
+    if (line == r"s SATISFIABLE") or (line == r"s UNSATISFIABLE"):
         status = SolverStatus.SUCCESS
         break
-    elif line == r's UNKNOWN':
+    elif line == r"s UNKNOWN":
         status = SolverStatus.TIMEOUT
         break
 
 outdir = {"status": status.value,
           "quality": 0,
-          "solver_call": solver_cmd + params}
+          "solver_call": solver_cmd}
 
 print(outdir)