+++ Examples/Resources/Solvers/PbO-CCSAT-Generic/_sparkle_solver_wrapper.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python3
 # -*- coding: UTF-8 -*-
-"""CSCCSat Solver wrapper."""
+"""Example solver wrapper for PbO-CCSAT in Python."""
 import sys
 import subprocess
 from pathlib import Path
 from sparkle.types import SolverStatus
-from sparkle.tools.solver_wrapper_parsing import parse_solver_wrapper_args
+from sparkle.tools.solver_wrapper_parsing import parse_solver_wrapper_args, \
+    get_solver_call_params
 
 
 # Parse the arguments of the solver wrapper
@@ -13,20 +14,24 @@ args_dict = parse_solver_wrapper_args(sys.argv[1:])
 
 # Extract certain args from the above dict for use further below
 solver_dir = args_dict["solver_dir"]
+objectives = args_dict["objectives"]
 instance = args_dict["instance"]
 seed = args_dict["seed"]
 
 # Construct the base solver call
-solver_name = "CSCCSat"
-if solver_dir != Path("."):
-    solver_exec = f"{solver_dir / solver_name}"
-else:
-    f"./{solver_name}"
-solver_cmd = [solver_exec, str(instance), str(seed)]
+solver_name = "PbO-CCSAT"
+solver_exec = f"{solver_dir / solver_name}" if solver_dir != Path(".") else "./" + \
+    solver_name
+solver_cmd = [solver_exec,
+              "-inst", str(instance),
+              "-seed", str(seed)]
 
-# CSCCSat does not have any configurable parameters, thus other params will not be added
+# Extract the parameters for the solver call
+params = get_solver_call_params(args_dict)
+
+# Execute the solver call
 try:
-    solver_call = subprocess.run(solver_cmd,
+    solver_call = subprocess.run(solver_cmd + params,
                                  capture_output=True)
 except Exception as ex:
     print(f"Solver call failed with exception:\n{ex}")
@@ -35,19 +40,21 @@ except Exception as ex:
 output_str = solver_call.stdout.decode()
 print(output_str)  # Print original output so it can be verified
 
-# Try to parse the status from the output
 status = SolverStatus.CRASHED
-for line in reversed(output_str.splitlines()):
+for line in output_str.splitlines():
     line = line.strip()
-    if (line == r"s SATISFIABLE") or (line == r"s UNSATISFIABLE"):
-        status = SolverStatus.SUCCESS
+    if (line == r's SATISFIABLE'):
+        status = SolverStatus.SAT
         break
-    elif line == r"s UNKNOWN":
+    elif (line == r's UNSATISFIABLE'):
+        status = SolverStatus.UNSAT
+        break
+    elif line == r's UNKNOWN':
         status = SolverStatus.TIMEOUT
         break
 
 outdir = {"status": status.value,
           "quality": 0,
-          "solver_call": solver_cmd}
+          "solver_call": solver_cmd + params}
 
 print(outdir)